+++ config/routes.rb
@@ -1,4 +1,5 @@
 Rails.application.routes.draw do
+  devise_for :users
   mount_avo
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html